Use `jf -` to read template from stdin
Also refactor the library, separating the CLI and template parsing
logic.

NOTE: We can either pass template or pass values from stdin, but not
both.
