+++ README.md
@@ -46,6 +46,18 @@ nix-env -f https://github.com/NixOS/nixpkgs/tarball/nixos-unstable -iA jf
 jf [OPTION]... [--] TEMPLATE [VALUE]... [NAME=VALUE]... [NAME@FILE]...
 ```
 
+### OPTIONS
+
+| option        | help                                            |
+| ------------- | ----------------------------------------------- |
+| -             | pass template via stdin                         |
+| --            | stop parsing CLI options                        |
+| -r, --raw     | print the raw rendered value without formatting |
+| -p, --pretty  | pretty print the JSON formatted output          |
+| -y, --yaml    | print the output as YAML instead of JSON        |
+| -h, --help    | print this help message                         |
+| -v, --version | print the version number                        |
+
 ### TEMPLATE
 
 A template is a string that should render into valid YAML. It can contain the
